+++ Assets/!Script/Player&Enemy/EnemyOnly/AttackSet.cs
@@ -15,9 +15,9 @@ public class AttackSet
     public int shootAngle ;
     [Header("Å‰‚Ì”­Ë‚Ü‚Å‚Ì’x‰„(•b)")]
     public int initialDelay = 0;
- public float startHP = 0f;
-    public float endHP = 0f;
-    public int damage = 1;
+    public int startHP = (int)0f;
+    public int endHP = (int)0f;
+    public float damage = 1;
     [Header("Pool‚·‚é—Ê")]
     public int poolSize = 30;
